fix: change repaint calls to update
Replaced `repaint` calls in capturewidget to use `update` instead, since
the change is not time sensitive. The main benefit from doing this is
that prior to this change, resizing the selection triggers a repaint
each frame from the `updateSizeIndicator` function in the callback in
`initSelection`, and the resulting output/animation isn't smooth, on my
PC at least.

This change improves the performance to be way smoother, subjectively.